refactor: resolve duplicated utils packages
