+++ pkg/utils/fixtures/karpenter/test_successful_ec2_node_class.tpl
