+++ pkg/utils/fixtures/test_multiple_template_b.tpl
