+++ pkg/utils/karpenter_utils.go
