+++ pkg/utils/karpenter_utils_test.go
