+++ pkg/utils/templates/ec2nodeclass.yaml.tpl
