+++ pkg/utils/templates/karpenter_custom_addon_boostrap.tf.tpl
