+++ pkg/utils/templates/launch_template_override.tf.tpl
