+++ pkg/utils/templates/spotinst_launch_spec_override.tf.tpl
