+++ pkg/utils/templates/spotinst_ocean_aws_override.tf.tpl
