+++ pkg/utils/templates/tests/data/karpenter_resource_output_node_pool_and_provisioner.yaml
@@ -102,7 +102,6 @@ spec:
       startupTaints:
       - effect: NoSchedule
         key: node.cloudprovider.kubernetes.io/uninitialized
-status: {}
 ---
 apiVersion: karpenter.k8s.aws/v1
 kind: EC2NodeClass
