+++ pkg/utils/templates/tests/data/karpenter_resource_output_node_pool_v1.yaml
@@ -53,7 +53,6 @@ spec:
       startupTaints:
       - effect: NoSchedule
         key: node.cloudprovider.kubernetes.io/uninitialized
-status: {}
 ---
 apiVersion: karpenter.k8s.aws/v1
 kind: EC2NodeClass
