+++ pkg/utils/templates/tests/data/karpenter_resource_output_provisioner.yaml
