+++ pkg/utils/templates/tests/karpenter_custom_addon_boostrap.tf
