+++ pkg/utils/templates/tests/launch_template_override.tf
