+++ pkg/utils/templates/tests/spotinst_launch_spec_override.tf
