+++ pkg/utils/templates/tests/spotinst_ocean_aws_override.tf
